Consolidate and cleanup login handling code
 - move OpenID, OAuth2, and user pass handling into one place
 - inject OAuth2 providers
 - general cleanup
