+++ modules/core/app/auth/oauth2/OAuth2Config.scala
@@ -1,0 +1,7 @@
+package auth.oauth2
+
+import auth.oauth2.providers.OAuth2Provider
+
+trait OAuth2Config {
+  def providers: Seq[OAuth2Provider]
+}
