+++ modules/portal/app/views/account/login.scala.html
@@ -1,4 +1,4 @@
-@(pwForm: Form[(String,String)], registerForm: Form[SignupData], registerAction: Call, recaptchaKey: String, openIdForm: Form[String], oauthProviders: Seq[auth.oauth2.providers.OAuth2Provider] = Seq.empty, isLogin: Boolean = true)(implicit userOpt: Option[UserProfile], request: RequestHeader, globalConfig: global.GlobalConfig, flash: Flash, messages: Messages,prefs: utils.SessionPrefs)
+@(pwForm: Form[(String,String)], registerForm: Form[SignupData], registerAction: Call, recaptchaKey: String, openIdForm: Form[String], oauthConfig: auth.oauth2.OAuth2Config, isLogin: Boolean = true)(implicit userOpt: Option[UserProfile], request: RequestHeader, globalConfig: global.GlobalConfig, flash: Flash, messages: Messages,prefs: utils.SessionPrefs)
 
 @implicitField = @{ views.html.helper.FieldConstructor(views.html.helpers.fieldTemplateSignUp.f) }
 
@@ -36,11 +36,11 @@
 
         <div class="tab-content">
             <div role="tabpanel" class="tab-pane @{if(isLogin) "active"}" id="login">
-                @oauth2Form("login.loginWith", openIdForm, oauthProviders, isLogin)
+                @oauth2Form("login.loginWith", openIdForm, oauthConfig, isLogin)
                 @loginForm(pwForm)
             </div>
             <div role="tabpanel" class="tab-pane @{if(!isLogin) "active"}" id="register">
-                @oauth2Form("login.signupWith", openIdForm, oauthProviders, isLogin)
+                @oauth2Form("login.signupWith", openIdForm, oauthConfig, isLogin)
                 @signupForm(registerForm, registerAction, recaptchaKey)
             </div>
         </div>
